requirements.txt
langchain
fastapi
uvicorn
openai
chromadb
faiss-cpu
python-dotenv
